Update treasurehunter.json
Data taken from
https://hypixel-skyblock.fandom.com/wiki/Treasure_Hunter#Treasure.
I have verified the changes on all lines (as they were updated by me on
the wiki).
The removed line 3 was a duplicate of the previous line 23,
the removed line 4 was likely an erroneous duplicate of the previous line 15,
and the removed line 5 was likely a duplicate of the previous line 24.